issue #158 feat: Compose Theme (DailyFilmTheme) 통일
