--- app/src/main/java/com/boostcamp/dailyfilm/presentation/searchfilm/ui/theme/Color.kt
@@ -1,30 +1,0 @@
-package com.boostcamp.dailyfilm.presentation.searchfilm.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val black = Color.Black
-val blackBlur = Color(0x4D000000)
-val white = Color.White
-val lightBlack = Color(0xFF202022)
-val lightGray = Color(0xFFE1E1E1)
-
-val primary = lightBlack
-val primaryVariant = white
-val background = white
-val surface = white
-val error = Color(0xFFB00020)
-val onPrimary = black
-val onBackground = black
-val onSurface = black
-val onError = white
-
-// darkTheme
-val darkPrimary = white
-val darkPrimaryVariant = lightBlack
-val darkBackground = lightBlack
-val darkSurface = Color(0xFF121212)
-val darkError = Color(0xFFCF6679)
-val darkOnPrimary = white
-val darkOnBackground = white
-val darkOnSurface = white
-val darkOnError = black
