+++ app/src/main/java/com/boostcamp/dailyfilm/presentation/ui/theme/Color.kt
@@ -1,19 +1,30 @@
-package com.example.composeapplication.ui.theme
+package com.boostcamp.dailyfilm.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val primary = Color(0xFF202022)
-val onPrimary = Color(0xFFFFFFFF)
-val surface = Color(0xFFFFFFFF)
-val onSurface = Color(0xFF202022)
-
-
-// nights
-val darkPrimary = Color(0xFFFFFFFF)
-val darkOnPrimary = Color(0xFFFFFFFF)
-val darkSurface = Color(0xFF202022)
-val darkOnSurface = Color(0xFFFFFFFF)
-
-val white = Color(0xFFFFFFFF)
+val black = Color.Black
+val blackBlur = Color(0x4D000000)
+val white = Color.White
 val lightBlack = Color(0xFF202022)
-val black = Color(0xFF000000)
+val lightGray = Color(0xFFE1E1E1)
+
+val primary = lightBlack
+val primaryVariant = white
+val background = white
+val surface = white
+val error = Color(0xFFB00020)
+val onPrimary = black
+val onBackground = black
+val onSurface = black
+val onError = white
+
+// darkTheme
+val darkPrimary = white
+val darkPrimaryVariant = lightBlack
+val darkBackground = lightBlack
+val darkSurface = Color(0xFF121212)
+val darkError = Color(0xFFCF6679)
+val darkOnPrimary = white
+val darkOnBackground = white
+val darkOnSurface = white
+val darkOnError = black
